requirements.txt
## Requirements without Version ##
django
djangorestframework
